Move version-related types into versions reducer
